+++ src/BrickApp/Handle.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE OverloadedStrings #-}
 -- | Brick application event handlers, depending on the current `RenderMode`, as used
 -- by this Brick application's `Brick.Main.appHandleEvent`.
 module BrickApp.Handle where
@@ -9,9 +10,11 @@ import qualified Brick.Types                   as B
 import qualified Graphics.Vty                  as V
 
 import BrickApp.Utils                           ( cacheRemove )
+import BrickApp.Utils.Popup                     ( popupDialogEventHandler)
 import BrickApp.Types.Names
 import BrickApp.Types.Helpers
 import BrickApp.Types
+import BrickApp.ModeAction.Homepage             ( goHome, createHomeDialog )
 import BrickApp.ModeAction.Help
 import BrickApp.ModeAction.Goto
 import BrickApp.ModeAction.Progress
@@ -48,17 +51,20 @@ appropriateHandler gbs e = case gbsRenderMode gbs of
   -- to the appropriate mode's handler)
   ProgressMode -> progressEventHandler gbs (Right e)
 
--- | Will do an event if the current `RenderMode` is in the allow list,
--- otherwise another event will be performed.
-doEventIfModes
+-- | Preform the successEvent if the function produces
+-- True when applied to the current RenderMode.
+--
+-- >>> doEventIfModeTrue gbs (`elem` [GotoMode, OpenConfigMode]) ...
+doEventIfModeTrue
   :: GopherBrowserState
-  -> [RenderMode]
+  -> (RenderMode -> Bool)
   -> B.EventM AnyName (B.Next GopherBrowserState)
   -> B.EventM AnyName (B.Next GopherBrowserState)
   -> B.EventM AnyName (B.Next GopherBrowserState)
-doEventIfModes gbs renderModesAllowList successEvent failEvent
-  | gbsRenderMode gbs `elem` renderModesAllowList = successEvent
+doEventIfModeTrue gbs func successEvent failEvent
+  | func $ gbsRenderMode gbs = successEvent
   | otherwise = failEvent
+
 
 -- FIXME: shouldn't history be handled top level and not in individual handlers? or are there
 -- some cases where we don't want history available
@@ -75,29 +81,42 @@ appEvent gbs (B.VtyEvent (V.EvKey (V.KChar 'q') [V.MCtrl])) = B.halt gbs
 appEvent gbs (B.VtyEvent e@(V.EvKey (V.KFun 5) [])) =
   let newCache = cacheRemove (gbsLocation gbs) (gbsCache gbs)
       newGbs   = gbs { gbsCache = newCache }
-  in  doEventIfModes gbs [TextFileMode, MenuMode] (liftIO (initProgressMode newGbs (Just $ gbsHistory gbs) (gbsLocation gbs)) >>= B.continue) (appropriateHandler gbs e)
+  in  doEventIfModeTrue gbs (`elem` [TextFileMode, MenuMode]) (liftIO (initProgressMode newGbs (Just $ gbsHistory gbs) (gbsLocation gbs)) >>= B.continue) (appropriateHandler gbs e)
+-- popup logic catching
+appEvent gbs@(GopherBrowserState{gbsPopup=(Just n)}) e = popupDialogEventHandler gbs n e
 -- Close a popup if there is one, otherwise forward to appropriate handler!
 appEvent gbs (B.VtyEvent e@(V.EvKey V.KEsc []))
   | hasPopup gbs = B.continue $ closePopup gbs
   | otherwise   = appropriateHandler gbs e
 -- add new bookmark
 appEvent gbs (B.VtyEvent e@(V.EvKey (V.KChar '+') [])) =
-  doEventIfModes gbs [TextFileMode, MenuMode] (B.continue $ initAddBookmarkMode gbs) (appropriateHandler gbs e)
+  doEventIfModeTrue gbs (`elem` [TextFileMode, MenuMode]) (B.continue $ initAddBookmarkMode gbs) (appropriateHandler gbs e)
 -- FIXME
 -- This is the config mode, which currently just goes right into the menu item
 -- command association editor.
 appEvent gbs (B.VtyEvent e@(V.EvKey (V.KChar 'c') [V.MCtrl])) =
   -- Why not just have this function defer to the appropriateHandler on failure?
-  doEventIfModes gbs [OpenConfigMode] (openConfigEventHandler gbs e) (initConfigOpenMode gbs)
--- Bookmark mode! FIXME this is a hack atm
+  doEventIfModeTrue gbs (== OpenConfigMode) (openConfigEventHandler gbs e) (liftIO (initConfigOpenMode gbs) >>= B.continue)
+-- Bookmark mode!
 appEvent gbs (B.VtyEvent (V.EvKey (V.KChar 'b') [V.MCtrl])) =
-  liftIO (initBookmarksMode gbs) >>= B.continue
--- `GotoMode`... should only activate if in `HelpMode`, `TextFileMode`, or `MenuMode`.
+  let successEvent = liftIO (initBookmarksMode gbs) >>= B.continue
+      failureEvent = B.continue gbs
+  in doEventIfModeTrue gbs (/= BookmarksMode) successEvent failureEvent
+-- GotoMode
 appEvent gbs (B.VtyEvent e@(V.EvKey (V.KChar 'g') [V.MCtrl])) =
-  doEventIfModes gbs [MenuMode, TextFileMode, HelpMode] (B.continue $ initGotoMode gbs) (appropriateHandler gbs e)
+  doEventIfModeTrue gbs (`elem` [MenuMode, TextFileMode, HelpMode]) (B.continue $ initGotoMode gbs) (appropriateHandler gbs e)
 -- TODO: needs to reset viewport
 appEvent gbs (B.VtyEvent e@(V.EvKey (V.KChar '?') [])) =
-  doEventIfModes gbs [HelpMode] (appropriateHandler gbs e) (liftIO (modifyGbsForHelp gbs) >>= B.continue)
+  doEventIfModeTrue gbs (== HelpMode) (appropriateHandler gbs e) (liftIO (modifyGbsForHelp gbs) >>= B.continue)
+-- Go to the homepage
+appEvent gbs (B.VtyEvent (V.EvKey (V.KChar 'h') [])) =
+  (liftIO $ goHome gbs) >>= B.continue
+-- Set the homepage
+-- FIXME: why does ctrl h not work?
+appEvent gbs (B.VtyEvent (V.EvKey (V.KChar 'z') [V.MCtrl])) =
+  -- TODO/FIXME: bring up prompt about setting homepage
+  B.continue (createHomeDialog gbs)
+--
 -- FIXME: this could be easily fixed just by doing appEvent gbs e instead of vtyevent
 -- and leaving it up to eventhandlers
 -- What about above FIXME... event types should be deicphered by event handler?