+++ src/BrickApp/ModeAction/Goto.hs
@@ -3,18 +3,17 @@
 -- | The heavy-lifting/actual actions of `GotoMode`.
 module BrickApp.ModeAction.Goto where
 
+import           Control.Monad.IO.Class
 import           Data.Maybe
 import qualified Data.Text                     as T
 
-import           Brick.Widgets.Core             ( txt )
+import qualified Brick.Types                   as B
+import qualified Brick.Main                    as B
 import           Brick.Widgets.Edit            as E
-import           Network.URI
 
 import BrickApp.Types
 import BrickApp.Types.Names
-import BrickApp.Utils
-import BrickApp.ModeAction.Progress
-import Gopher
+import BrickApp.Utils.WaffleAddresses
 
 initGotoMode :: GopherBrowserState -> GopherBrowserState
 initGotoMode gbs = gbs
@@ -31,47 +30,16 @@ initGotoMode gbs = gbs
 -- FIXME: what if bad input?! what if can't resolve? errors in network need better handling
 -- FIXME: what if NOT a menu!
 -- should this be a part of progress? or ge called by progress instead?
-mkGotoResponseState :: GopherBrowserState -> IO GopherBrowserState
+--mkGotoResponseState :: GopherBrowserState -> IO GopherBrowserState
+mkGotoResponseState :: GopherBrowserState -> B.EventM AnyName (B.Next GopherBrowserState)
 mkGotoResponseState gbs =
   -- get the host, port, selector
   let unparsedURI = T.filter (/= '\n')
         $ T.unlines (E.getEditContents $ seEditorState $ fromJust $ gbsStatus gbs)
       formerGbs = formerMode gbs
-  in  either (errorPopup gbs unparsedURI) (initProgressMode formerGbs Nothing) (tryLocationOrFail unparsedURI)
- where
-  prefixSchemeIfMissing :: T.Text -> T.Text
-  prefixSchemeIfMissing potentialURI
-    | "gopher://" `T.isPrefixOf` potentialURI = potentialURI
-    | otherwise                               = "gopher://" <> potentialURI
-
-  errorPopup :: GopherBrowserState -> T.Text -> T.Text -> IO GopherBrowserState
-  errorPopup gbs' someBadURI message =
-    let pop = Popup
-                { pLabel   = "Goto input error!"
-                , pWidgets = [txt message]
-                , pHelp    = "Invalid:" <> someBadURI
-                }
-    in  pure $ gbs' { gbsPopup = Just pop }
-
-  -- | Try to parse a `Location` from `Text` (which is hopefully
-  -- some kind of valid URI), or give back an error message.
-  tryLocationOrFail :: T.Text -> Either T.Text (T.Text, Int, T.Text, RenderMode, Maybe T.Text)
-  tryLocationOrFail potentialURI = do
-    parsedURI <- case (parseURI . T.unpack $ prefixSchemeIfMissing potentialURI) of
-      Just uri -> Right uri
-      Nothing  -> Left "Cannot even begin to parse supplied URI!"
-    authority' <- case uriAuthority parsedURI of
-      Just auth -> Right auth
-      Nothing   -> Left $ "Invalid URI (no authority)."
-    regName    <- case uriRegName authority' of
-      ""      -> Left "Invalid URI (no regname/host)."
-      rn      -> Right rn
-    port <- case uriPort authority' of
-      ""     -> Right 70
-      ':':p  -> Right (read p :: Int)
-      _      -> Left $ "Invalid URI (bad port)." -- I don' think this ever can occur with Network.URI...
-    let resource = uriPath parsedURI
-    Right (T.pack regName, port, removeGopherType $ T.pack resource, selectorToRenderMode $ T.pack resource, Nothing)
+  -- here is where i detect type first
+  -- I should modularize this to be used elsewhere like home or follow links?
+  in  liftIO (loadAddress formerGbs unparsedURI Nothing) >>= B.continue
 
 -- | Revert to mode prior to `GotoMode` being initiated.
 formerMode :: GopherBrowserState -> GopherBrowserState